Reorganize modules into a filesystem hierarchy
