+++ hpccm/cmake.py
@@ -25,9 +25,9 @@ import logging # pylint: disable=unused-import
 import os
 import re
 
-from hpccm.comment import comment
 from hpccm.packages import packages
-from hpccm.shell import shell
+from hpccm.primitives.comment import comment
+from hpccm.primitives.shell import shell
 from hpccm.wget import wget
 
 class cmake(wget):
