+++ hpccm/mkl.py
@@ -27,12 +27,12 @@ import re
 
 import hpccm.config
 
-from hpccm.comment import comment
 from hpccm.common import linux_distro
-from hpccm.copy import copy
-from hpccm.environment import environment
 from hpccm.packages import packages
-from hpccm.shell import shell
+from hpccm.primitives.comment import comment
+from hpccm.primitives.copy import copy
+from hpccm.primitives.environment import environment
+from hpccm.primitives.shell import shell
 from hpccm.toolchain import toolchain
 from hpccm.wget import wget
 from hpccm.yum import yum
