+++ hpccm/mvapich2.py
@@ -28,14 +28,14 @@ from copy import copy as _copy
 
 import hpccm.config
 
-from hpccm.comment import comment
 from hpccm.common import linux_distro
-from hpccm.ConfigureMake import ConfigureMake
-from hpccm.copy import copy
-from hpccm.environment import environment
 from hpccm.packages import packages
+from hpccm.ConfigureMake import ConfigureMake
+from hpccm.primitives.comment import comment
+from hpccm.primitives.copy import copy
+from hpccm.primitives.environment import environment
+from hpccm.primitives.shell import shell
 from hpccm.sed import sed
-from hpccm.shell import shell
 from hpccm.tar import tar
 from hpccm.toolchain import toolchain
 from hpccm.wget import wget
