+++ hpccm/ofed.py
@@ -22,8 +22,8 @@ from __future__ import print_function
 
 import logging # pylint: disable=unused-import
 
-from hpccm.comment import comment
 from hpccm.packages import packages
+from hpccm.primitives.comment import comment
 
 class ofed(object):
     """OFED building block"""
