+++ hpccm/pgi.py
@@ -26,12 +26,12 @@ import os
 
 import hpccm.config
 
-from hpccm.comment import comment
 from hpccm.common import linux_distro
-from hpccm.copy import copy
-from hpccm.environment import environment
 from hpccm.packages import packages
-from hpccm.shell import shell
+from hpccm.primitives.comment import comment
+from hpccm.primitives.copy import copy
+from hpccm.primitives.environment import environment
+from hpccm.primitives.shell import shell
 from hpccm.tar import tar
 from hpccm.toolchain import toolchain
 from hpccm.wget import wget
