+++ hpccm/primitives/blob.py
