+++ hpccm/python.py
@@ -23,8 +23,8 @@ from __future__ import print_function
 
 import logging # pylint: disable=unused-import
 
-from hpccm.comment import comment
 from hpccm.packages import packages
+from hpccm.primitives.comment import comment
 
 class python(object):
     """Python building block"""
